Release API Security

+++ src/main/java/cz/cvut/fit/ejk/gaidumax/drive/controller/FileController.java
@@ -4,6 +4,8 @@ import cz.cvut.fit.ejk.gaidumax.drive.dto.FileDto;
 import cz.cvut.fit.ejk.gaidumax.drive.dto.UpdateFileDto;
 import cz.cvut.fit.ejk.gaidumax.drive.mapper.FileMapper;
 import cz.cvut.fit.ejk.gaidumax.drive.service.interfaces.FileService;
+import cz.cvut.fit.ejk.gaidumax.drive.service.security.interfaces.AuthService;
+import io.quarkus.security.Authenticated;
 import jakarta.inject.Inject;
 import jakarta.validation.Valid;
 import jakarta.ws.rs.Consumes;
@@ -16,6 +18,8 @@ import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 
+import java.util.UUID;
+
 @Path("/files")
 public class FileController {
 
@@ -23,12 +27,16 @@ public class FileController {
     FileService fileService;
     @Inject
     FileMapper fileMapper;
+    @Inject
+    AuthService authService;
 
     @Deprecated(forRemoval = true)
     @GET
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public FileDto get(@PathParam("id") Long id) {
+    @Authenticated
+    public FileDto get(@PathParam("id") UUID id) {
+        authService.checkUserHasAccessToFile(id);
         var file = fileService.getByIdOrThrow(id);
         return fileMapper.toDto(file);
     }
@@ -36,6 +44,7 @@ public class FileController {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
+    @Authenticated
     public FileDto create(@Valid FileDto fileDto) {
         var file = fileService.create(fileDto);
         return fileMapper.toDto(file);
@@ -45,14 +54,18 @@ public class FileController {
     @Path("/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public FileDto update(@PathParam("id") Long id, @Valid UpdateFileDto fileDto) {
+    @Authenticated
+    public FileDto update(@PathParam("id") UUID id, @Valid UpdateFileDto fileDto) {
+        authService.checkUserHasAccessToFile(id);
         var file = fileService.update(id, fileDto);
         return fileMapper.toDto(file);
     }
 
     @DELETE
     @Path("/{id}")
-    public void update(@PathParam("id") Long id) {
+    @Authenticated
+    public void update(@PathParam("id") UUID id) {
+        authService.checkUserHasAccessToFile(id);
         fileService.delete(id);
     }
 }
